% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignatedMedicalEvents.R
\docType{package}
\name{DesignatedMedicalEvents-package}
\alias{DesignatedMedicalEvents}
\alias{DesignatedMedicalEvents-package}
\title{DesignatedMedicalEvents: A Study Package for Generating Cohort Diagnostics}
\description{
A package that generates cohort diagnostics for cohort definitions included in the package.
}
\keyword{internal}
